requirements.txt
fastapi
uvicorn[standard]

sqlalchemy>=2.0
alembic

psycopg2-binary

pydantic[email]

python-jose
passlib[bcrypt]==1.7.4
python-dotenv

python-multipart

pytest
httpx